% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importAndProcess.R
\name{importCharacter}
\alias{importCharacter}
\title{Import character from google drive}
\usage{
importCharacter(regex = NULL, fileID = NULL, output = NULL,
  overwrite = TRUE)
}
\arguments{
\item{regex}{A regular expression that matches the file name. If there are multiple matches, the most recently edited file will the chosen}

\item{fileID}{A googledrive file ID. If regex is not provided}

\item{output}{if provided, the file will be saved here. if not, a temporary file will be used}

\item{overwrite}{If TRUE the new file will overwrite the old one. Not important if output isn't provided as the file will be deleted eventually}
}
\value{
The output is a nested list extracted from the original XML file, edited to make certain fields more computer readable.
}
\description{
Provided a regex or fileID this will import the character from your google drive so it can be used in utility functions. Will ask for
googledrive authentication on first load. If you are using Rstudio server, you can authenticate using another machine and move the generated
.httr-oauth file to your working directory.
}
