% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{weaponBonus}
\alias{weaponBonus}
\title{Calculate bonuses that affect a weapon attack}
\usage{
weaponBonus(weapon, char = getOption("defaultCharacter"))
}
\description{
Calculate bonuses that affect a weapon attack
}
