% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{attack}
\alias{attack}
\title{Attack}
\usage{
attack(adv = 0, sharpShoot = F, attackStat = c("Str", "Dex", "Con", "Wis",
  "Int", "Chr"), damageDice = "1d6", proficient = TRUE, modToHit = 0,
  modToDamage = 0, ammo = NULL, vocal = TRUE, critRange = 20,
  char = getOption("defaultCharacter"))
}
\arguments{
\item{adv}{integer. 0: Normal, -1 Disadvantage, +1 Advantage}

\item{sharpShoot}{logical. are we sharpshooting}

\item{attackStat}{character. which stat to use for attack. 3 letter. Added to hit and damage rolls}

\item{damageDice}{character vector. attack dice. If more than one dice type is used (as in sneak attack) add them as a vector (\code{damageDice = c('1d8','1d6')})}

\item{proficient}{logical. is the attacker proficient in the weapon. If so proficiency bonus will be added to hit rolls}

\item{modToHit}{integer. hit modifier}

\item{modToDamage}{integer. damage modifier}

\item{ammo}{character. If provided ammo will be used. Before using, you need to set your ammo count manually by character$ammoName = 30 (character$bolt = 30)}

\item{vocal}{logical. Should dice rolls and crit notifications be printed.}

\item{critRange}{vector of integers. Which rolls to count as a critical hit}

\item{char}{A list outputted from importCharacter or processCharacter, or a character naming such a list}
}
\value{
Named vector. Attack roll and damage. If a 1 is rolled damage is always 0
}
\description{
Attack
}

